Orquestrator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orquestrator"
version = "1.11.0"
description = "Agente Orquestador KYB — coordina Colorado, Arizona y Nevada (persiste directamente en PostgreSQL)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "httpx>=0.27.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "asyncpg>=0.30.0",
    "tenacity>=8.2.0",
]
